{{cookiecutter.project_name}}/requirements.txt
-e .
-e git+https://github.com/django/django.git@1.6b4#egg=Django
dj_database_url==0.2.1
boto==2.9.5
django-storages==1.1.8
django-braces==1.2.2
South==0.8.2
django-model-utils==1.5.0
Pillow==2.2.1

newrelic
